Extract schema merge into doltlite_schema_merge.c, rename tests, add 26 tests
Move migrateSchemaRowData and helpers from doltlite.c (1400+ lines)
into a dedicated doltlite_schema_merge.c. Rename doltlite_alter_merge.sh
to doltlite_schema_merge.sh.

Add 7 new test sections (26 assertions):
- Both branches add different columns with data → merged correctly
- Convergent column additions (identical) → no conflict
- Type conflicts detected and errored
- One-sided column + data changes propagated
- Multiple tables with independent schema changes
- Data integrity with 50 rows + column + diff-based migration
